MB-9804 Add `crpytography` to our `requirements-dev.txt`
It's needed for the test utils from `locust`. Unfortunately `locust` is still using a deprecated way (`tests_require`) of defining their requirements for tests, which means we can't import them more easily (e.g. by listing it as `locust[test]` or something. Maybe in a future version we'll be able to do that rather than listing an individual requirement.