MB-10533 Set up possible alternative for forming URLs
This keeps us from having to define this at a user level, which is good because a given user will likely have to make requests of other subdomains, e.g. the prime tasks make calls to the internal API. That's the whole reason MilMoveRequestMixin looks the way it does. I'll have to keep thinking on whether this is actually cleaner though.