MB-10533 Update the docs on running tests
Add more clarity around unit tests vs load tests.
Move relevant information from https://github.com/transcom/milmove_load_testing/wiki/Running-Load-Tests-Against-MyMove to this README.md. We phased out the wiki in the `mymove` repo and this page is the only one we have in this repo, so maybe we should just get rid of it.
The table of contents has now been updated with the pre-commit hook instead of the old script we used.