MB-10533 Update `LoginTaskSet` to use `RestMixin`
Can't add `MilMoveRequestMixin` because it clashes with the `MilMoveHostMixin` being used by the user classes that currently use the `LoginTaskSet`. Once we update those users to stop using the old mixin, then `LoginTaskSet` can be updated to just subclass `RestTaskSet` fully.
Also add an example of using the `LoginTaskSet` to the `locustfiles/rest.py`.