MB-10534 Add a few load tests
In writing the example for expected failures in the readme, I ended up with some load tests that we didn't fully have, so I figured I should add them.
We had previously had a `list_moves` load test that existed to set the default mto IDs, but we got rid of it when we started creating the move in the `PrimeTasks.on_start` method. I agree we don't need it for that, but realized that we probably should still have load tests for the endpoint itself.