Add dest address & save locator when creating shipment
We could also add the destination address as the prime later on
but I think, since the addresses depend on each other, the load testing
code is simpler when they're defined in the same place.

I also added the move locator to the flow context, which will help
office users find the move in their queues.